+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [main] Don't panic when parsing options. Instead list valid values and exit.
 - [main] `--alsa-mixer-device` and `--alsa-mixer-index` now fallback to the card and index specified in `--device`.
 - [core] Removed unsafe code (breaking)
+- [playback] Adhere to ReplayGain spec when calculating gain normalisation factor.
 
 ### Removed
 - [playback] `alsamixer`: previously deprecated option `mixer-card` has been removed.